src/sql-tools/requirements.txt
openai==1.37.0
langchain-core==0.2.22
pydantic==2.8.2
python-dotenv==1.0.1
tiktoken==0.7.0
azure-search-documents==11.5.0
azure-identity==1.17.1
uvicorn==0.30.3
streamlit==1.36.0
promptflow==1.13.0
promptflow-core==1.13.0
langchain-community==0.2.9
langchain-openai==0.1.17
opentelemetry-instrumentation-langchain==0.25.5
langchain==0.2.10
langgraph==0.1.9
langchain-text-splitters==0.2.2
langchain-qdrant==0.1.2
bs4==0.0.2
langchainhub==0.1.20
langchain_experimental==0.0.63
langchain_community==0.2.9
sqlalchemy==2.0.31
pyodbc==5.1.0